Implemented `toJS` in such a way that it only recurses into observables
